SE: Deduplicate arithmetic calculations (#7938)
