+++ analyzers/tests/SonarAnalyzer.UnitTest/SymbolicExecution/Roslyn/RoslynSymbolicExecutionTest.CompoundAssignment.cs
@@ -121,598 +121,4 @@ public partial class RoslynSymbolicExecutionTest
         validator.TagValue("C").Should().HaveOnlyConstraint(ObjectConstraint.NotNull);
         validator.TagValue("Result").Should().HaveOnlyConstraint(ObjectConstraint.NotNull);
     }
-
-    [DataTestMethod]
-    [DataRow("+=", 43)]
-    [DataRow("-=", 41)]
-    [DataRow("*=", 42)]
-    [DataRow("/=", 42)]
-    [DataRow("%=",  0)]
-    [DataRow("&=",  0)]
-    [DataRow("|=", 43)]
-    [DataRow("^=", 43)]
-    public void CompoundAssignment_Arithmetics_Int(string op, int expected)
-    {
-        var code = $"""
-            var value = 42;
-            var result = value {op} 1;
-            Tag("Result", result);
-            Tag("Value", value);
-            """;
-        var validator = SETestContext.CreateCS(code).Validator;
-        validator.TagValue("Result").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-    }
-
-    [DataTestMethod]
-    [DataRow("<<=")]
-    [DataRow(">>=")]
-    [DataRow(">>>=")]
-    public void CompoundAssignment_Arithmetics_Int_ShiftOperators(string op)
-    {
-        var code = $"""
-            var value = 42;
-            var result = value {op} 1;
-            Tag("Result", result);
-            Tag("Value", value);
-            """;
-        var validator = SETestContext.CreateCS(code).Validator;
-        validator.TagValue("Result").Should().HaveOnlyConstraints(ObjectConstraint.NotNull);
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull);
-    }
-
-    [DataTestMethod]
-    [DataRow("+=")]
-    [DataRow("-=")]
-    [DataRow("*=")]
-    [DataRow("/=")]
-    [DataRow("%=")]
-    [DataRow("&=")]
-    [DataRow("|=")]
-    [DataRow("^=")]
-    [DataRow("<<=")]
-    [DataRow(">>=")]
-    [DataRow(">>>=")]
-    public void CompoundAssignment_Arithmetics_Char(string op)
-    {
-        var code = $"""
-            var value = 'z';
-            var result = value {op} 'a';
-            Tag("Result", result);
-            Tag("Value", value);
-            """;
-        var validator = SETestContext.CreateCS(code).Validator;
-        validator.TagValue("Result").Should().HaveOnlyConstraints(ObjectConstraint.NotNull);
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull);
-    }
-
-    [DataTestMethod]
-    [DataRow("&=")]
-    [DataRow("|=")]
-    [DataRow("^=")]
-    public void CompoundAssignment_Logical(string op)
-    {
-        var code = $"""
-            var value = true;
-            var result = value {op} true;
-            Tag("Result", result);
-            Tag("Value", value);
-            """;
-        var validator = SETestContext.CreateCS(code).Validator;
-        validator.TagValue("Result").Should().HaveOnlyConstraints(ObjectConstraint.NotNull);
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull);
-    }
-
-    [DataTestMethod]
-    [DataRow("i += j", 43)]
-    [DataRow("i += 2", 44)]
-    [DataRow("i += (true ? 1 : 0)", 43)]
-    [DataRow("i -= j", 41)]
-    [DataRow("i -= 2", 40)]
-    [DataRow("i -= (true ? 1 : 0)", 41)]
-    public void CompoundAssignment_Arithmetic_PlusAndMinus(string expression, int expected)
-    {
-        var code = $"""
-            var i = 42;
-            var j = 1;
-            var result = {expression};
-            Tag("Result", result);
-            Tag("I", i);
-            """;
-        var validator = SETestContext.CreateCS(code).Validator;
-        validator.TagValue("Result").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-        validator.TagValue("I").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-    }
-
-    [DataTestMethod]
-    [DataRow(5, 0, 0)]
-    [DataRow(5, 1, 5)]
-    [DataRow(5, -1, -5)]
-    [DataRow(5, 4, 20)]
-    [DataRow(5, -4, -20)]
-    [DataRow(-5, 0, 0)]
-    [DataRow(-5, 1, -5)]
-    [DataRow(-5, -1, 5)]
-    [DataRow(-5, 4, -20)]
-    [DataRow(-5, -4, 20)]
-    public void CompoundAssignment_Multiplication_SingleValue(int left, int right, int expected)
-    {
-        var code = $"""
-            var left = {left};
-            var right = {right};
-            var value = left *= right;
-            Tag("Value", value);
-            Tag("Left", left);
-            """;
-        var validator = SETestContext.CreateCS(code).Validator;
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-        validator.TagValue("Left").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-    }
-
-    [DataTestMethod]
-    [DataRow("i >=  3 && j >=  5", 15, null)]
-    [DataRow("i >=  3 && j >= -5", null, null)]
-    [DataRow("i >=  3 && j <=  5", null, null)]
-    [DataRow("i >=  3 && j <= -5", null, -15)]
-    [DataRow("i >= -3 && j >=  5", null, null)]
-    [DataRow("i >= -3 && j >= -5", null, null)]
-    [DataRow("i >= -3 && j <=  5", null, null)]
-    [DataRow("i >= -3 && j <= -5", null, null)]
-    [DataRow("i <=  3 && j >=  5", null, null)]
-    [DataRow("i <=  3 && j >= -5", null, null)]
-    [DataRow("i <=  3 && j <=  5", null, null)]
-    [DataRow("i <=  3 && j <= -5", null, null)]
-    [DataRow("i <= -3 && j >=  5", null, -15)]
-    [DataRow("i <= -3 && j >= -5", null, null)]
-    [DataRow("i <= -3 && j <=  5", null, null)]
-    [DataRow("i <= -3 && j <= -5", 15, null)]
-    [DataRow("i ==  3 && j >=  5", 15, null)]
-    [DataRow("i ==  3 && j >= -5", -15, null)]
-    [DataRow("i ==  3 && j <=  5", null, 15)]
-    [DataRow("i ==  3 && j <= -5", null, -15)]
-    [DataRow("i == -3 && j >=  5", null, -15)]
-    [DataRow("i == -3 && j >= -5", null, 15)]
-    [DataRow("i == -3 && j <=  5", -15, null)]
-    [DataRow("i == -3 && j <= -5", 15, null)]
-    public void CompoundAssignment_Multiplication_Range(string expression, int? expectedMin, int? expectedMax)
-    {
-        var code = $$"""
-            if ({{expression}})
-            {
-                var value = i *= j;
-                Tag("Value", value);
-                Tag("I", i);
-            }
-            """;
-        var validator = SETestContext.CreateCS(code, "int i, int j").Validator;
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-        validator.TagValue("I").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-    }
-
-    [DataTestMethod]
-    [DataRow(21, 4, 5)]
-    [DataRow(21, -4, -5)]
-    [DataRow(-21, 4, -5)]
-    [DataRow(-21, -4, 5)]
-    [DataRow(4, 5, 0)]
-    [DataRow(-4, 5, 0)]
-    [DataRow(4, -5, 0)]
-    [DataRow(-4, -5, 0)]
-    [DataRow(5, 1, 5)]
-    [DataRow(5, -1, -5)]
-    [DataRow(-5, 1, -5)]
-    [DataRow(-5, -1, 5)]
-    public void CompoundAssignment_Division_SingleValue(int left, int right, int expected)
-    {
-        var code = $"""
-            var left = {left};
-            var right = {right};
-            var value = left /= right;
-            Tag("Value", value);
-            Tag("Left", left);
-            """;
-        var validator = SETestContext.CreateCS(code).Validator;
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-        validator.TagValue("Left").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-    }
-
-    [DataTestMethod]
-    [DataRow("i >=  21 && j >=  5", 0, null)]
-    [DataRow("i >=  21 && j >= -5", null, null)]
-    [DataRow("i >=  21 && j ==  5", 4, null)]
-    [DataRow("i >=  21 && j == -5", null, -4)]
-    [DataRow("i >=  21 && j <=  5", null, null)]
-    [DataRow("i >=  21 && j <= -5", null, 0)]
-    [DataRow("i >= -21 && j >=  5", -4, null)]
-    [DataRow("i >= -21 && j >= -5", null, null)]
-    [DataRow("i >= -21 && j ==  5", -4, null)]
-    [DataRow("i >= -21 && j == -5", null, 4)]
-    [DataRow("i >= -21 && j <=  5", null, null)]
-    [DataRow("i >= -21 && j <= -5", null, 4)]
-    [DataRow("i ==  21 && j >=  5", 0, 4)]
-    [DataRow("i ==  21 && j >= -5", -21, 21)]
-    [DataRow("i ==  21 && j <=  5", -21, 21)]
-    [DataRow("i ==  21 && j <= -5", -4, 0)]
-    [DataRow("i == -21 && j >=  5", -4, 0)]
-    [DataRow("i == -21 && j >= -5", -21, 21)]
-    [DataRow("i == -21 && j <=  5", -21, 21)]
-    [DataRow("i == -21 && j <= -5", 0, 4)]
-    [DataRow("i <=  21 && j >=  5", null, 4)]
-    [DataRow("i <=  21 && j >= -5", null, null)]
-    [DataRow("i <=  21 && j ==  5", null, 4)]
-    [DataRow("i <=  21 && j == -5", -4, null)]
-    [DataRow("i <=  21 && j <=  5", null, null)]
-    [DataRow("i <=  21 && j <= -5", -4, null)]
-    [DataRow("i <= -21 && j >=  5", null, 0)]
-    [DataRow("i <= -21 && j >= -5", null, null)]
-    [DataRow("i <= -21 && j ==  5", null, -4)]
-    [DataRow("i <= -21 && j == -5", 4, null)]
-    [DataRow("i <= -21 && j <=  5", null, null)]
-    [DataRow("i <= -21 && j <= -5", 0, null)]
-    [DataRow("i >= -21 && i <= 15 && j >= -5", -21, 21)]
-    [DataRow("i >= -15 && i <= 21 && j >= -5", -21, 21)]
-    [DataRow("i >=  21 && j >= 0", 0, null)]
-    [DataRow("i >=  21 && j == 0", null, null)]
-    [DataRow("i >=  21 && j <= 0", null, 0)]
-    [DataRow("i >= -21 && j >= 0", -21, null)]
-    [DataRow("i >= -21 && j == 0", null, null)]
-    [DataRow("i >= -21 && j <= 0", null, 21)]
-    [DataRow("i <=  21 && j >= 0", null, 21)]
-    [DataRow("i <=  21 && j == 0", null, null)]
-    [DataRow("i <=  21 && j <= 0", -21, null)]
-    [DataRow("i <= -21 && j >= 0", null, 0)]
-    [DataRow("i <= -21 && j == 0", null, null)]
-    [DataRow("i <= -21 && j <= 0", 0, null)]
-    public void CompoundAssignment_Division_Range(string expression, int? expectedMin, int? expectedMax)
-    {
-        var code = $$"""
-            if ({{expression}})
-            {
-                var value = i /= j;
-                Tag("Value", value);
-                Tag("I", i);
-            }
-            """;
-        var validator = SETestContext.CreateCS(code, "int i, int j").Validator;
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-        validator.TagValue("I").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-    }
-
-    [DataTestMethod]
-    [DataRow(21, 4, 1)]
-    [DataRow(21, -4, 1)]
-    [DataRow(-21, 4, -1)]
-    [DataRow(-21, -4, -1)]
-    [DataRow(4, 5, 4)]
-    [DataRow(-4, 5, -4)]
-    [DataRow(4, -5, 4)]
-    [DataRow(-4, -5, -4)]
-    [DataRow(5, 5, 0)]
-    [DataRow(5, -5, 0)]
-    [DataRow(-5, 5, 0)]
-    [DataRow(-5, -5, 0)]
-    [DataRow(5, 1, 0)]
-    [DataRow(5, -1, 0)]
-    [DataRow(-5, 1, 0)]
-    [DataRow(-5, -1, 0)]
-    public void CompoundAssignment_Remainder_SingleValue(int left, int right, int expected)
-    {
-        var code = $"""
-            var left = {left};
-            var right = {right};
-            var value = left %= right;
-            Tag("Value", value);
-            Tag("Left", left);
-            """;
-        var validator = SETestContext.CreateCS(code).Validator;
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-        validator.TagValue("Left").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-    }
-
-    [DataTestMethod]
-    [DataRow("i >=  21 && j >=  5", 0, null)]
-    [DataRow("i >=  21 && j >= -5", 0, null)]
-    [DataRow("i >=  21 && j ==  5", 0, 4)]
-    [DataRow("i >=  21 && j == -5", 0, 4)]
-    [DataRow("i >=  21 && j <=  5", 0, null)]
-    [DataRow("i >=  21 && j <= -5", 0, null)]
-    [DataRow("i >= -21 && j >=  5", -21, null)]
-    [DataRow("i >= -21 && j >= -5", -21, null)]
-    [DataRow("i >= -21 && j ==  5", -4, 4)]
-    [DataRow("i >= -21 && j == -5", -4, 4)]
-    [DataRow("i >= -21 && j <=  5", -21, null)]
-    [DataRow("i >= -21 && j <= -5", -21, null)]
-    [DataRow("i ==  21 && j >=  5", 0, 21)]
-    [DataRow("i ==  21 && j >= -5", 0, 21)]
-    [DataRow("i ==  21 && j <=  5", 0, 21)]
-    [DataRow("i ==  21 && j <= -5", 0, 21)]
-    [DataRow("i == -21 && j >=  5", -21, 0)]
-    [DataRow("i == -21 && j >= -5", -21, 0)]
-    [DataRow("i == -21 && j <=  5", -21, 0)]
-    [DataRow("i == -21 && j <= -5", -21, 0)]
-    [DataRow("i <=  21 && j >=  5", null, 21)]
-    [DataRow("i <=  21 && j >= -5", null, 21)]
-    [DataRow("i <=  21 && j ==  5", -4, 4)]
-    [DataRow("i <=  21 && j == -5", -4, 4)]
-    [DataRow("i <=  21 && j <=  5", null, 21)]
-    [DataRow("i <=  21 && j <= -5", null, 21)]
-    [DataRow("i <= -21 && j >=  5", null, 0)]
-    [DataRow("i <= -21 && j >= -5", null, 0)]
-    [DataRow("i <= -21 && j ==  5", -4, 0)]
-    [DataRow("i <= -21 && j == -5", -4, 0)]
-    [DataRow("i <= -21 && j <=  5", null, 0)]
-    [DataRow("i <= -21 && j <= -5", null, 0)]
-    [DataRow("i >=  21 && j >= 0", 0, null)]
-    [DataRow("i >=  21 && j == 0", null, null)]
-    [DataRow("i >=  21 && j <= 0", 0, null)]
-    [DataRow("i >= -21 && j >= 0", -21, null)]
-    [DataRow("i >= -21 && j == 0", null, null)]
-    [DataRow("i >= -21 && j <= 0", -21, null)]
-    [DataRow("i <=  21 && j >= 0", null, 21)]
-    [DataRow("i <=  21 && j == 0", null, null)]
-    [DataRow("i <=  21 && j <= 0", null, 21)]
-    [DataRow("i <= -21 && j >= 0", null, 0)]
-    [DataRow("i <= -21 && j == 0", null, null)]
-    [DataRow("i <= -21 && j <= 0", null, 0)]
-    [DataRow("i >=  21 && j == 1", 0, 0)]
-    [DataRow("i >= -21 && j == 1", 0, 0)]
-    [DataRow("i <=  21 && j == 1", 0, 0)]
-    [DataRow("i <= -21 && j == 1", 0, 0)]
-    [DataRow("i >=  21 && j >=   5 && j <= 10", 0, 9)]
-    [DataRow("i >=  21 && j >=  -5 && j <= 10", 0, 9)]
-    [DataRow("i >=  21 && j >= -10 && j <=  5", 0, 9)]
-    [DataRow("i >=  21 && j >= -10 && j <= -5", 0, 9)]
-    [DataRow("i >= -21 && j >=   5 && j <= 10", -9, 9)]
-    [DataRow("i >= -21 && j >=  -5 && j <= 10", -9, 9)]
-    [DataRow("i >= -21 && j >= -10 && j <=  5", -9, 9)]
-    [DataRow("i >= -21 && j >= -10 && j <= -5", -9, 9)]
-    [DataRow("i <=  21 && j >=   5 && j <= 10", -9, 9)]
-    [DataRow("i <=  21 && j >=  -5 && j <= 10", -9, 9)]
-    [DataRow("i <=  21 && j >= -10 && j <=  5", -9, 9)]
-    [DataRow("i <=  21 && j >= -10 && j <= -5", -9, 9)]
-    [DataRow("i <= -21 && j >=   5 && j <= 10", -9, 0)]
-    [DataRow("i <= -21 && j >=  -5 && j <= 10", -9, 0)]
-    [DataRow("i <= -21 && j >= -10 && j <=  5", -9, 0)]
-    [DataRow("i <= -21 && j >= -10 && j <= -5", -9, 0)]
-    [DataRow("i >=   5 && i <=  10 && j >=  21", 5, 10)]
-    [DataRow("i >=   5 && i <=  10 && j <= -21", 5, 10)]
-    [DataRow("i >=  -5 && i <=  10 && j >=  21", -5, 10)]
-    [DataRow("i >=  -5 && i <=  10 && j <= -21", -5, 10)]
-    [DataRow("i >= -10 && i <=   5 && j >=  21", -10, 5)]
-    [DataRow("i >= -10 && i <=   5 && j <= -21", -10, 5)]
-    [DataRow("i >= -10 && i <=  -5 && j >=  21", -10, -5)]
-    [DataRow("i >= -10 && i <=  -5 && j <= -21", -10, -5)]
-    [DataRow("i >=   4 && i <=   7 && j >=   5 && j <= 10", 0, 7)]
-    [DataRow("i == 5 && j == 0", null, null)]
-    public void CompoundAssignment_Remainder_Range(string expression, int? expectedMin, int? expectedMax)
-    {
-        var code = $$"""
-            if ({{expression}})
-            {
-                var value = i %= j;
-                Tag("Value", value);
-                Tag("I", i);
-            }
-            """;
-        var validator = SETestContext.CreateCS(code, "int i, int j").Validator;
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-        validator.TagValue("I").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-    }
-
-    [DataTestMethod]
-    [DataRow(0b0000, 0b0000, 0b0000)]
-    [DataRow(0b0101, 0b0101, 0b0101)]
-    [DataRow(0b0101, 0b0001, 0b0001)]
-    [DataRow(0b1010, 0b0110, 0b0010)]
-    [DataRow(0b1010, 0b0000, 0b0000)]
-    [DataRow(0b1111, 0b1111, 0b1111)]
-    [DataRow(5, -5, 1)]
-    [DataRow(5, -4, 4)]
-    [DataRow(-5, -5, -5)]
-    [DataRow(-5, -4, -8)]
-    public void CompoundAssignment_BitAnd_SingleValue(int left, int right, int expected)
-    {
-        var code = $"""
-            var left = {left};
-            var right = {right};
-            var value = left &= right;
-            Tag("Value", value);
-            Tag("Left", left);
-            """;
-        var validator = SETestContext.CreateCS(code).Validator;
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-        validator.TagValue("Left").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-    }
-
-    [DataTestMethod]
-    [DataRow("i >=  3 && j >=  5", 0, null)]
-    [DataRow("i >=  3 && j >= -5", 0, null)]
-    [DataRow("i >=  3 && j <=  5", 0, null)]
-    [DataRow("i >=  3 && j <= -5", 0, null)]
-    [DataRow("i >= -3 && j >=  5", 0, null)]
-    [DataRow("i >= -3 && j >= -5", -8, null)]
-    [DataRow("i >= -3 && j <=  5", null, null)]
-    [DataRow("i >= -3 && j <= -5", null, null)]
-    [DataRow("i ==  3 && j >=  5", 0, 3)]
-    [DataRow("i ==  3 && j >= -5", 0, 3)]
-    [DataRow("i ==  3 && j <=  5", 0, 3)]
-    [DataRow("i ==  3 && j <= -5", 0, 3)]
-    [DataRow("i == -3 && j >=  5", 0, null)]
-    [DataRow("i == -3 && j >= -5", -8, null)]
-    [DataRow("i == -3 && j <=  5", null, 5)]
-    [DataRow("i == -3 && j <= -5", null, -5)]
-    [DataRow("i <=  3 && j >=  5", 0, null)]
-    [DataRow("i <=  3 && j >= -5", null, null)]
-    [DataRow("i <=  3 && j <=  5", null, 5)]
-    [DataRow("i <=  3 && j <= -5", null, 3)]
-    [DataRow("i <= -3 && j >=  5", 0, null)]
-    [DataRow("i <= -3 && j >= -5", null, null)]
-    [DataRow("i <= -3 && j <=  5", null, 5)]
-    [DataRow("i <= -3 && j <= -5", null, -5)]
-    public void CompoundAssignment_BitAnd_Range(string expression, int? expectedMin, int? expectedMax)
-    {
-        var code = $$"""
-            if ({{expression}})
-            {
-                var value = i &= j;
-                Tag("Value", value);
-                Tag("I", i);
-            }
-            """;
-        var validator = SETestContext.CreateCS(code, "int i, int j").Validator;
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-        validator.TagValue("I").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-    }
-
-    [DataTestMethod]
-    [DataRow(0b0000, 0b0000, 0b0000)]
-    [DataRow(0b0101, 0b0101, 0b0101)]
-    [DataRow(0b0101, 0b0001, 0b0101)]
-    [DataRow(0b1010, 0b0110, 0b1110)]
-    [DataRow(0b1010, 0b0000, 0b1010)]
-    [DataRow(0b1111, 0b1111, 0b1111)]
-    [DataRow(5, -5, -1)]
-    [DataRow(5, -4, -3)]
-    [DataRow(-5, -5, -5)]
-    [DataRow(-5, -4, -1)]
-    public void CompoundAssignment_BitOr_SingleValue(int left, int right, int expected)
-    {
-        var code = $"""
-            var left = {left};
-            var right = {right};
-            var value = left |= right;
-            Tag("Value", value);
-            Tag("Left", left);
-            """;
-        var validator = SETestContext.CreateCS(code).Validator;
-        validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-        validator.TagValue("Left").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-    }
-
-    [DataTestMethod]
-    [DataRow("i >=  4 && j >=  6", 6, null)]
-    [DataRow("i >=  4 && j >= -6", -6, null)]
-    [DataRow("i >=  4 && j <=  6", null, null)]
-    [DataRow("i >=  4 && j <= -6", null, -1)]
-    [DataRow("i >= -4 && j >=  6", -4, null)]
-    [DataRow("i >= -4 && j >= -6", -6, null)]
-    [DataRow("i >= -4 && j <=  6", null, null)]
-    [DataRow("i >= -4 && j <= -6", null, -1)]
-    [DataRow("i ==  4 && j >=  6", 6, null)]
-    [DataRow("i ==  4 && j >= -6", -6, null)] // (i | j) >= -4, we only get a lower bound of min
-    [DataRow("i ==  4 && j <=  6", null, 7)]
-    [DataRow("i ==  4 && j <= -6", null, -1)] // (i | j) <= -2, we only get an upper bound of max
-    [DataRow("i == -4 && j >=  6", -4, -1)]
-    [DataRow("i == -4 && j >= -6", -4, -1)]
-    [DataRow("i == -4 && j <=  6", -4, -1)]
-    [DataRow("i == -4 && j <= -6", -4, -1)]
-    [DataRow("i <=  4 && j >=  6", null, null)]
-    [DataRow("i <=  4 && j >= -6", null, null)]
-    [DataRow("i <=  4 && j <=  6", null, 7)]
-    [DataRow("i <=  4 && j <= -6", null, -1)]
-    [DataRow("i <= -4 && j >=  6", null, -1)]
-    [DataRow("i <= -4 && j >= -6", null, -1)]
-    [DataRow("i <= -4 && j <=  6", null, -1)]
-    [DataRow("i <= -4 && j <= -6", null, -1)]
-    public void CompoundAssignment_BitOr_Range(string expression, int? expectedMin, int? expectedMax)
-    {
-        var code = $$"""
-            if ({{expression}})
-            {
-                var value = i |= j;
-                Tag("Value", value);
-                Tag("I", i);
-            }
-            """;
-
-        var validator = SETestContext.CreateCS(code, "int i, int j").Validator;
-        if (expectedMin is not null || expectedMax is not null)
-        {
-            validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-            validator.TagValue("I").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-        }
-        else
-        {
-            validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull);
-        }
-    }
-
-    [DataTestMethod]
-    [DataRow(0b0000, 0b0000, 0b0000)]
-    [DataRow(0b0101, 0b0101, 0b0000)]
-    [DataRow(0b0101, 0b0001, 0b0100)]
-    [DataRow(0b1010, 0b0110, 0b1100)]
-    [DataRow(0b1010, 0b0000, 0b1010)]
-    [DataRow(0b1111, 0b1111, 0b0000)]
-    [DataRow(5, -5, -2)]
-    [DataRow(5, -4, -7)]
-    [DataRow(-5, -5, 0)]
-    [DataRow(-5, -4, 7)]
-    public void CompoundAssignment_BitXor_SingleValue(int left, int right, int expected)
-    {
-        var code = $"""
-            var left = {left};
-            var right = {right};
-            var value = left ^= right;
-            Tag("Value", value);
-            Tag("Left", left);
-            """;
-        SETestContext.CreateCS(code).Validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expected));
-    }
-
-    [DataTestMethod]
-    [DataRow("i >=  4 && j >=  6", 0, null)]
-    [DataRow("i >=  4 && j >= -6", null, null)]
-    [DataRow("i >=  4 && j <=  6", null, null)]
-    [DataRow("i >=  4 && j <= -6", null, -1)]
-    [DataRow("i >= -4 && j >=  6", null, null)]
-    [DataRow("i >= -4 && j >= -6", null, null)]
-    [DataRow("i >= -4 && j <=  6", null, null)]
-    [DataRow("i >= -4 && j <= -6", null, null)]  // exact range: null, -1
-    [DataRow("i ==  4 && j >=  6", 2, null)]
-    [DataRow("i ==  4 && j >= -6", -8, null)]
-    [DataRow("i ==  4 && j <=  6", null, 7)]
-    [DataRow("i ==  4 && j <= -6", null, -1)]    // exact range: null, -2
-    [DataRow("i == -4 && j >=  6", null, -1)]    // exact range: null, -5
-    [DataRow("i == -4 && j >= -6", null, 7)]
-    [DataRow("i == -4 && j <=  6", -8, null)]
-    [DataRow("i == -4 && j <= -6", 2, null)]     // exact range: 4, null
-    [DataRow("i <=  4 && j >=  6", null, null)]
-    [DataRow("i <=  4 && j >= -6", null, null)]
-    [DataRow("i <=  4 && j <=  6", null, null)]
-    [DataRow("i <=  4 && j <= -6", null, null)]
-    [DataRow("i <= -4 && j >=  6", null, -1)]
-    [DataRow("i <= -4 && j >= -6", null, null)]
-    [DataRow("i <= -4 && j <=  6", null, null)]
-    [DataRow("i <= -4 && j <= -6", 0, null)]
-    [DataRow("i >=  4 && j >=  6 && j <=  8", 0, null)]
-    [DataRow("i >=  4 && j >=  1 && j <=  3", 1, null)]           // exact range: 4, null
-    [DataRow("i >=  4 && i <=  6 && j >=  6", 0, null)]
-    [DataRow("i >=  4 && i <=  6 && j >=  6 && j <= 8", 0, 15)]   // exact range: 0, 14
-    [DataRow("i >=  4 && i <=  5 && j >=  6 && j <= 8", 1, 15)]   // exact range: 2, 13
-    [DataRow("i >= -3 && i <= -1 && j >= -4 && j <=-2", 0, 7)]    // exact range: 0, 3
-    [DataRow("i >= -4 && i <= -3 && j >= -2 && j <=-1", 1, 7)]    // exact range: 2, 3
-    [DataRow("i >= -4 && i <=  6 && j >= -3 && j <= 8", -16, 15)] // exact range: -12, 14
-    public void CompoundAssignment_BitXor_Range(string expression, int? expectedMin, int? expectedMax)
-    {
-        var code = $$"""
-            if ({{expression}})
-            {
-                var value = i ^= j;
-                Tag("Value", value);
-                Tag("I", i);
-            }
-            """;
-
-        var validator = SETestContext.CreateCS(code, "int i, int j").Validator;
-        if (expectedMin is not null || expectedMax is not null)
-        {
-            validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-            validator.TagValue("I").Should().HaveOnlyConstraints(ObjectConstraint.NotNull, NumberConstraint.From(expectedMin, expectedMax));
-        }
-        else
-        {
-            validator.TagValue("Value").Should().HaveOnlyConstraints(ObjectConstraint.NotNull);
-        }
-    }
 }
